Content_Agent/pyproject.toml
[project]
name = "content-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[google-genai,tools]>=1.7.1",
    "exa-py>=2.0.1",
    "ipywidgets>=8.1.8",
    "python-dotenv>=1.1.1",
]
